+++ apps/codeimage/src/components/Toolbar/LanguageSelectorButton.tsx
@@ -1,5 +1,5 @@
 import {Popover, PopoverButton} from 'solid-headless';
-import {Button} from '@codeimage/ui';
+import {Button, FadeInOutTransition} from '@codeimage/ui';
 import {For} from 'solid-js';
 import {useFloating} from '../../core/floating-ui/floating-ui';
 import {offset} from '@floating-ui/dom';
@@ -7,7 +7,6 @@ import {useI18n} from '@codeimage/locale';
 import {DropdownItem} from '../../ui/Dropdown/DropdownItem';
 import {DropdownMenu} from '../../ui/Dropdown/DropdownMenu';
 import {AppLocaleEntries} from '../../i18n';
-import {FadeInOutTransition} from '../../ui/Transition/Transition';
 import {SvgIcon} from '../../ui/SvgIcon/SvgIcon';
 
 interface LanguageSelectorButtonProps {
