+++ apps/codeimage/src/ui/Toast/Snackbar.tsx
@@ -6,7 +6,7 @@ import {Dynamic} from 'solid-js/web';
 import {Button, Text} from '@codeimage/ui';
 import {Box} from '@codeimage/ui';
 import {SvgIcon} from '../SvgIcon/SvgIcon';
-import {FadeInOutWithScaleTransition} from '../Transition/Transition';
+import {FadeInOutWithScaleTransition} from '@codeimage/ui';
 
 export function SnackBar(props: SnackbarData & {id: string}): JSX.Element {
   const [isOpen, setIsOpen] = createSignal(true);
