+++ packages/ui/src/lib/primitives/Dialog/index.ts
@@ -1,0 +1,3 @@
+export * from './Dialog';
+export * from './DialogPanel';
+export * from './DialogTitle';
