+++ packages/ui/src/lib/primitives/Transition/index.ts
@@ -1,0 +1,1 @@
+export * from './Transition';
