+++ packages/ui/src/lib/theme/variables.css.ts
@@ -2,3 +2,4 @@ import {createVar} from '@vanilla-extract/css';
 
 export const colorVar = createVar();
 export const backgroundColorVar = createVar();
+export const dynamicFullHeight = createVar();
